Cleanup acceptance tester
