+++ tests/_support/Page/Acceptance/Administrator/AdminPage.php
@@ -51,6 +51,22 @@ class AdminPage extends \AcceptanceTester
 	public static $iconSearch = ['class' => 'icon-search'];
 
 	/**
+	 * Locator for the Tabs in Edit View
+	 *
+	 * @var    array
+	 * @since  __DEPLOY_VERSION__
+	 */
+	public static $tabsLocator = ['xpath' => "//ul[@id='myTabTabs']/li/a"];
+
+	/**
+	 * Locator for the Check All checkbox
+	 *
+	 * @var    array
+	 * @since  __DEPLOY_VERSION__
+	 */
+	public static $checkAll = ['xpath' => "//thead//input[@name='checkall-toggle' or @name='toggle']"];
+
+	/**
 	 * Method to search using user's name
 	 *
 	 * @param   string  $keyword  The keyword to search
@@ -116,5 +132,184 @@ class AdminPage extends \AcceptanceTester
 	{
 		$I = $this;
 		$I->waitForText($title, TIMEOUT, self::$pageTitle);
+	}
+
+	/**
+	 * Function to select Toolbar buttons in Joomla! Admin Toolbar Panel
+	 *
+	 * @param   string  $button  The full name of the button
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 *
+	 * @return  void
+	 */
+	public function clickToolbarButton($button)
+	{
+		$I     = $this;
+		$input = strtolower($button);
+
+		$suiteConfiguration = $I->getSuiteConfiguration();
+		$screenWidth        = explode("x", $suiteConfiguration['modules']['config']['JoomlaBrowser']['window_size']);
+
+		if ($screenWidth[0] <= 480)
+		{
+			$I->click('Toolbar');
+		}
+
+		switch ($input)
+		{
+			case "new":
+				$I->click(['xpath' => "//div[@id='toolbar-new']//button"]);
+				break;
+			case "edit":
+				$I->click(['xpath' => "//div[@id='toolbar-edit']//button"]);
+				break;
+			case "publish":
+				$I->click(['xpath' => "//div[@id='toolbar-publish']//button"]);
+				break;
+			case "unpublish":
+				$I->click(['xpath' => "//div[@id='toolbar-unpublish']//button"]);
+				break;
+			case "archive":
+				$I->click(['xpath' => "//div[@id='toolbar-archive']//button"]);
+				break;
+			case "check-in":
+				$I->click(['xpath' => "//div[@id='toolbar-checkin']//button"]);
+				break;
+			case "batch":
+				$I->click(['xpath' => "//div[@id='toolbar-batch']//button"]);
+				break;
+			case "rebuild":
+				$I->click(['xpath' => "//div[@id='toolbar-refresh']//button"]);
+				break;
+			case "trash":
+				$I->click(['xpath' => "//div[@id='toolbar-trash']//button"]);
+				break;
+			case "save":
+				$I->click(['xpath' => "//div[@id='toolbar-apply']//button"]);
+				break;
+			case "save & close":
+				$I->click(['xpath' => "//div[@id='toolbar-save']//button"]);
+				break;
+			case "save & new":
+				$I->click(['xpath' => "//div[@id='toolbar-save-new']//button"]);
+				break;
+			case "cancel":
+				$I->click(['xpath' => "//div[@id='toolbar-cancel']//button"]);
+				break;
+			case "options":
+				$I->click(['xpath' => "//div[@id='toolbar-options']//button"]);
+				break;
+			case "empty trash":
+			case "delete":
+				$I->click(['xpath' => "//div[@id='toolbar-delete']//button"]);
+				break;
+			case "unblock":
+				$I->click(['xpath' => "//div[@id='toolbar-unblock']//button"]);
+				break;
+			case "featured":
+				$I->click(['xpath' => "//div[@id='toolbar-featured']//button"]);
+				break;
+		}
+	}
+
+	/**
+	 * Function to select all the item in the Search results in Administrator List
+	 *
+	 * Note: We recommend use of checkAllResults function only after searchForItem to be sure you are selecting only the desired result set
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 *
+	 * @return  void
+	 */
+	public function checkAllResults()
+	{
+		$I = $this;
+
+		$I->comment("Selecting Checkall button");
+		$I->click(self::$checkAll);
+	}
+
+	/**
+	 * Selects an option in a Chosen Selector based on its id
+	 *
+	 * @param   string  $selectId  The id of the <select> element
+	 * @param   string  $option    The text in the <option> to be selected in the chosen selector
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 *
+	 * @return  void
+	 */
+	public function selectOptionInChosenById($selectId, $option)
+	{
+		$chosenSelectID = $selectId . '_chzn';
+
+		$I = $this;
+		$I->comment("I open the $chosenSelectID chosen selector");
+		$I->click(['xpath' => "//div[@id='$chosenSelectID']/a/div/b"]);
+		$I->comment("I select $option");
+		$I->click(['xpath' => "//div[@id='$chosenSelectID']//li[text()='$option']"]);
+
+		// Gives time to chosen to close
+		$I->wait(1);
+	}
+
+	/**
+	 * Function to Logout from Administrator Panel in Joomla!
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 *
+	 * @return  void
+	 */
+	public function doAdministratorLogout()
+	{
+		$I = $this;
+		$I->click(
+			['xpath' => "//ul[@class='nav nav-user pull-right']//li//a[@class='dropdown-toggle']"]
+		);
+
+		$I->comment("I click on Top Right corner toggle to Logout from Admin");
+		$I->waitForElement(
+			['xpath' => "//li[@class='dropdown open']/ul[@class='dropdown-menu']//a[text() = 'Logout']"],
+			TIMEOUT
+		);
+
+		$I->click(
+			['xpath' => "//li[@class='dropdown open']/ul[@class='dropdown-menu']//a[text() = 'Logout']"]
+		);
+
+		$I->waitForElement(['id' => 'mod-login-username'], TIMEOUT);
+		$I->waitForText(
+			'Log in',
+			TIMEOUT,
+			['xpath' => "//fieldset[@class='loginform']//button"]
+		);
+	}
+
+	/**
+	 * Function to Verify the Tabs on a Joomla! screen
+	 *
+	 * @param   array  $expectedTabs  Expected Tabs on the Page
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 *
+	 * @return  void
+	 */
+	public function verifyAvailableTabs($expectedTabs)
+	{
+		$I = $this;
+
+		$actualArrayOfTabs = $I->grabMultiple(self::$tabsLocator);
+
+		$I->comment(
+			"Fetch the current list of Tabs in the edit view which is: " . implode(", ", $actualArrayOfTabs)
+		);
+
+		$I->assertEquals(
+			$expectedTabs,
+			$actualArrayOfTabs, "Tab Labels do not match on edit view of" . $I->grabFromCurrentUrl()
+		);
+
+		$I->comment('Verify the Tabs');
 	}
 }
